refactor: extract helper to get environment id
